LayoutTests/fast/loader/scroll-position-restored-on-reload-at-load-event.html

<!DOCTYPE html>
<html>
  <head>
    <script src="../../resources/js-test.js"></script>
    <script>
    description('Test ensures that scrollingElement.scrollTop/Left properties are available by the time DOMContentLoaded event fires.');
    // Navigation steps:
    // 1- page gets first loaded and scrolled.
    // 2- reload is performed.
    // Test: ensure that by the time DOMContenLoaded fires (after a reload navigation), scrollingElement.scrollTop/Left are set.

    function init(evt) {
        if (window.name == 'second/load') {
            shouldBe('document.scrollingElement.scrollTop', '2000');
            shouldBe('document.scrollingElement.scrollLeft', '1000');
            window.name = "";

            if (window.testRunner)
                finishJSTest();
        } else {
            window.scrollTo(1000, 2000);

            window.name = "second/load";
            setTimeout('location.reload(true)', 0);
        }
    }

    window.addEventListener('DOMContentLoaded', init, true);
    window.onunload = function() {}  // prevent caching

    var jsTestIsAsync = true;
    </script>
    <body>
        <div style='width: 9999px; height:9999px; float:left;'></div>
        <h1 id='console'/>
    </body>
</html>
